backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
xgboost
tensorflow
torch
requests
matplotlib
seaborn
joblib
gunicorn
faker
imbalanced-learn
pyyaml
user-agents
flask-bcrypt
pyjwt
flask-sqlalchemy
